+++ beacon_node/beacon_chain/src/kzg_utils.rs
@@ -1,7 +1,7 @@
 use kzg::{Blob as KzgBlob, Bytes48, Cell as KzgCell, Error as KzgError, Kzg};
 use std::sync::Arc;
-use types::data_column_sidecar::{Cell, DataColumn};
-use types::{Blob, DataColumnSidecar, EthSpec, Hash256, KzgCommitment, KzgProof, KzgProofs};
+use types::data_column_sidecar::Cell;
+use types::{Blob, DataColumnSidecar, EthSpec, Hash256, KzgCommitment, KzgProof};
 
 /// Converts a blob ssz List object to an array to be used with the kzg
 /// crypto library.
@@ -71,102 +71,6 @@ where
     kzg.verify_cell_proof_batch(&cells, &proofs, &coordinates, &commitments)
 }
 
-pub fn reconstruct_data_columns<E: EthSpec>(
-    kzg: &Kzg,
-    data_columns: &[Arc<DataColumnSidecar<E>>],
-) -> Result<Vec<Arc<DataColumnSidecar<E>>>, KzgError> {
-    let mut columns = vec![Vec::with_capacity(E::max_blobs_per_block()); E::number_of_columns()];
-    let mut column_kzg_proofs =
-        vec![Vec::with_capacity(E::max_blobs_per_block()); E::number_of_columns()];
-
-    for col in 0..E::number_of_columns() {
-        let mut cells: Vec<KzgCell> = vec![];
-        let mut cell_ids: Vec<u64> = vec![];
-        for data_column in data_columns {
-            let cell = data_column
-                .column
-                .get(col)
-                .ok_or(KzgError::InconsistentArrayLength(format!(
-                    "Missing data column at index {col}"
-                )))?;
-
-            cells.push(ssz_cell_to_crypto_cell::<E>(cell)?);
-            cell_ids.push(data_column.index);
-        }
-        // recover_all_cells does not expect sorted
-        let all_cells = kzg.recover_all_cells(&cell_ids, &cells)?;
-        let blob = kzg.cells_to_blob(&all_cells)?;
-        // Note: This function computes all cells and proofs. According to Justin this is okay,
-        // computing a partial set may be more expensive and requires code paths that don't exist.
-        // Computing the blobs cells is technically unnecessary but very cheap. It's done here again
-        // for simplicity.
-        let (blob_cells, blob_cell_proofs) = kzg.compute_cells_and_proofs(&blob)?;
-
-        // we iterate over each column, and we construct the column from "top to bottom",
-        // pushing on the cell and the corresponding proof at each column index. we do this for
-        // each blob (i.e. the outer loop).
-        let cell = blob_cells
-            .get(col)
-            .ok_or(KzgError::InconsistentArrayLength(format!(
-                "Missing blob cell at index {col}"
-            )))?;
-        let cell: Vec<u8> = cell
-            .into_inner()
-            .into_iter()
-            .flat_map(|data| (*data).into_iter())
-            .collect();
-        let cell = Cell::<E>::from(cell);
-
-        let proof = blob_cell_proofs
-            .get(col)
-            .ok_or(KzgError::InconsistentArrayLength(format!(
-                "Missing blob cell KZG proof at index {col}"
-            )))?;
-
-        let column = columns
-            .get_mut(col)
-            .ok_or(KzgError::InconsistentArrayLength(format!(
-                "Missing data column at index {col}"
-            )))?;
-        let column_proofs =
-            column_kzg_proofs
-                .get_mut(col)
-                .ok_or(KzgError::InconsistentArrayLength(format!(
-                    "Missing data column proofs at index {col}"
-                )))?;
-
-        column.push(cell);
-        column_proofs.push(*proof);
-    }
-
-    // Clone sidecar elements from existing data column, no need to re-compute
-    let first_data_column = data_columns
-        .first()
-        .ok_or(KzgError::InconsistentArrayLength(
-            "data_columns should have at least one element".to_string(),
-        ))?;
-    let kzg_commitments = &first_data_column.kzg_commitments;
-    let signed_block_header = &first_data_column.signed_block_header;
-    let kzg_commitments_inclusion_proof = &first_data_column.kzg_commitments_inclusion_proof;
-
-    let sidecars: Vec<Arc<DataColumnSidecar<E>>> = columns
-        .into_iter()
-        .zip(column_kzg_proofs)
-        .enumerate()
-        .map(|(index, (col, proofs))| {
-            Arc::new(DataColumnSidecar {
-                index: index as u64,
-                column: DataColumn::<E>::from(col),
-                kzg_commitments: kzg_commitments.clone(),
-                kzg_proofs: KzgProofs::<E>::from(proofs),
-                signed_block_header: signed_block_header.clone(),
-                kzg_commitments_inclusion_proof: kzg_commitments_inclusion_proof.clone(),
-            })
-        })
-        .collect();
-    Ok(sidecars)
-}
-
 /// Validate a batch of blob-commitment-proof triplets from multiple `BlobSidecars`.
 pub fn validate_blobs<E: EthSpec>(
     kzg: &Kzg,
@@ -223,4 +127,63 @@ pub fn verify_kzg_proof<E: EthSpec>(
     y: Hash256,
 ) -> Result<bool, KzgError> {
     kzg.verify_kzg_proof(kzg_commitment, &z.0.into(), &y.0.into(), kzg_proof)
+}
+
+#[cfg(test)]
+mod test {
+    use bls::Signature;
+    use eth2_network_config::TRUSTED_SETUP_BYTES;
+    use kzg::{Kzg, KzgCommitment, TrustedSetup};
+    use types::{
+        beacon_block_body::KzgCommitments, BeaconBlock, BeaconBlockDeneb, Blob, BlobsList,
+        ChainSpec, DataColumnSidecar, EmptyBlock, EthSpec, MainnetEthSpec, SignedBeaconBlock,
+    };
+
+    #[test]
+    fn build_and_reconstruct() {
+        type E = MainnetEthSpec;
+        let num_of_blobs = 6;
+        let spec = E::default_spec();
+        let (signed_block, blob_sidecars) = create_test_block_and_blobs::<E>(num_of_blobs, &spec);
+
+        let trusted_setup: TrustedSetup = serde_json::from_reader(TRUSTED_SETUP_BYTES)
+            .map_err(|e| format!("Unable to read trusted setup file: {}", e))
+            .expect("should have trusted setup");
+        let kzg = Kzg::new_from_trusted_setup(trusted_setup).expect("should create kzg");
+
+        let column_sidecars =
+            DataColumnSidecar::build_sidecars(&blob_sidecars, &signed_block, &kzg).unwrap();
+
+        // Now reconstruct
+        let reconstructed_columns = DataColumnSidecar::reconstruct(
+            &kzg,
+            &column_sidecars.iter().as_slice()[0..column_sidecars.len() / 2],
+        )
+        .unwrap();
+
+        for i in 0..E::number_of_columns() {
+            assert_eq!(reconstructed_columns.get(i), column_sidecars.get(i), "{i}");
+        }
+    }
+
+    fn create_test_block_and_blobs<E: EthSpec>(
+        num_of_blobs: usize,
+        spec: &ChainSpec,
+    ) -> (SignedBeaconBlock<E>, BlobsList<E>) {
+        let mut block = BeaconBlock::Deneb(BeaconBlockDeneb::empty(spec));
+        let mut body = block.body_mut();
+        let blob_kzg_commitments = body.blob_kzg_commitments_mut().unwrap();
+        *blob_kzg_commitments =
+            KzgCommitments::<E>::new(vec![KzgCommitment::empty_for_testing(); num_of_blobs])
+                .unwrap();
+
+        let signed_block = SignedBeaconBlock::from_block(block, Signature::empty());
+
+        let blobs = (0..num_of_blobs)
+            .map(|_| Blob::<E>::default())
+            .collect::<Vec<_>>()
+            .into();
+
+        (signed_block, blobs)
+    }
 }
