+++ consensus/types/src/data_column_sidecar.rs
@@ -9,7 +9,7 @@ use bls::Signature;
 use derivative::Derivative;
 #[cfg_attr(test, double)]
 use kzg::Kzg;
-use kzg::{Blob as KzgBlob, Error as KzgError};
+use kzg::{Blob as KzgBlob, Cell as KzgCell, Error as KzgError};
 use kzg::{KzgCommitment, KzgProof};
 use merkle_proof::MerkleTreeError;
 #[cfg(test)]
@@ -169,6 +169,106 @@ impl<E: EthSpec> DataColumnSidecar<E> {
         Ok(sidecars)
     }
 
+    pub fn reconstruct(kzg: &Kzg, data_columns: &[Arc<Self>]) -> Result<Vec<Arc<Self>>, KzgError> {
+        let mut columns =
+            vec![Vec::with_capacity(E::max_blobs_per_block()); E::number_of_columns()];
+        let mut column_kzg_proofs =
+            vec![Vec::with_capacity(E::max_blobs_per_block()); E::number_of_columns()];
+
+        let first_data_column = data_columns
+            .first()
+            .ok_or(KzgError::InconsistentArrayLength(
+                "data_columns should have at least one element".to_string(),
+            ))?;
+        let num_of_blobs = first_data_column.kzg_commitments.len();
+
+        for row_index in 0..num_of_blobs {
+            let mut cells: Vec<KzgCell> = vec![];
+            let mut cell_ids: Vec<u64> = vec![];
+            for data_column in data_columns {
+                let cell =
+                    data_column
+                        .column
+                        .get(row_index)
+                        .ok_or(KzgError::InconsistentArrayLength(format!(
+                            "Missing data column at index {row_index}"
+                        )))?;
+
+                cells.push(ssz_cell_to_crypto_cell::<E>(cell)?);
+                cell_ids.push(data_column.index);
+            }
+            // recover_all_cells does not expect sorted
+            let all_cells = kzg.recover_all_cells(&cell_ids, &cells)?;
+            let blob = kzg.cells_to_blob(&all_cells)?;
+
+            // Note: This function computes all cells and proofs. According to Justin this is okay,
+            // computing a partial set may be more expensive and requires code paths that don't exist.
+            // Computing the blobs cells is technically unnecessary but very cheap. It's done here again
+            // for simplicity.
+            let (blob_cells, blob_cell_proofs) = kzg.compute_cells_and_proofs(&blob)?;
+
+            // we iterate over each column, and we construct the column from "top to bottom",
+            // pushing on the cell and the corresponding proof at each column index. we do this for
+            // each blob (i.e. the outer loop).
+            for col in 0..E::number_of_columns() {
+                let cell = blob_cells
+                    .get(col)
+                    .ok_or(KzgError::InconsistentArrayLength(format!(
+                        "Missing blob cell at index {col}"
+                    )))?;
+                let cell: Vec<u8> = cell
+                    .into_inner()
+                    .into_iter()
+                    .flat_map(|data| (*data).into_iter())
+                    .collect();
+                let cell = Cell::<E>::from(cell);
+
+                let proof = blob_cell_proofs
+                    .get(col)
+                    .ok_or(KzgError::InconsistentArrayLength(format!(
+                        "Missing blob cell KZG proof at index {col}"
+                    )))?;
+
+                let column = columns
+                    .get_mut(col)
+                    .ok_or(KzgError::InconsistentArrayLength(format!(
+                        "Missing data column at index {col}"
+                    )))?;
+                let column_proofs =
+                    column_kzg_proofs
+                        .get_mut(col)
+                        .ok_or(KzgError::InconsistentArrayLength(format!(
+                            "Missing data column proofs at index {col}"
+                        )))?;
+
+                column.push(cell);
+                column_proofs.push(*proof);
+            }
+        }
+
+        // Clone sidecar elements from existing data column, no need to re-compute
+        let kzg_commitments = &first_data_column.kzg_commitments;
+        let signed_block_header = &first_data_column.signed_block_header;
+        let kzg_commitments_inclusion_proof = &first_data_column.kzg_commitments_inclusion_proof;
+
+        let sidecars: Vec<Arc<DataColumnSidecar<E>>> = columns
+            .into_iter()
+            .zip(column_kzg_proofs)
+            .enumerate()
+            .map(|(index, (col, proofs))| {
+                Arc::new(DataColumnSidecar {
+                    index: index as u64,
+                    column: DataColumn::<E>::from(col),
+                    kzg_commitments: kzg_commitments.clone(),
+                    kzg_proofs: KzgProofs::<E>::from(proofs),
+                    signed_block_header: signed_block_header.clone(),
+                    kzg_commitments_inclusion_proof: kzg_commitments_inclusion_proof.clone(),
+                })
+            })
+            .collect();
+        Ok(sidecars)
+    }
+
     pub fn min_size() -> usize {
         // min size is one cell
         Self {
@@ -271,6 +371,12 @@ pub type DataColumnSidecarList<E> =
     VariableList<Arc<DataColumnSidecar<E>>, <E as EthSpec>::DataColumnCount>;
 pub type FixedDataColumnSidecarList<E> =
     FixedVector<Option<Arc<DataColumnSidecar<E>>>, <E as EthSpec>::DataColumnCount>;
+
+/// Converts a cell ssz List object to an array to be used with the kzg
+/// crypto library.
+fn ssz_cell_to_crypto_cell<E: EthSpec>(cell: &Cell<E>) -> Result<KzgCell, KzgError> {
+    KzgCell::from_bytes(cell.as_ref()).map_err(Into::into)
+}
 
 #[cfg(test)]
 mod test {
